# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BakeryStock API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /movements:
    get:
      operationId: listStockMovements
      tags: [movements]
      summary: List stock movements
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: itemId
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/MovementType"
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockMovement"
    post:
      operationId: createStockMovement
      tags: [movements]
      summary: Record a stock movement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStockMovementRequest"
      responses:
        "201":
          description: Movement recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"
  /movements/{id}:
    get:
      operationId: getStockMovement
      tags: [movements]
      summary: Get stock movement by ID
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Stock movement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockMovement"
...
  /reports/download:
    get:
      operationId: downloadReport
      tags: [reports]
      summary: Download a report as CSV or PDF
      parameters:
        - name: type
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/ReportType"
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [csv, pdf]
        - name: branchId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Report file
          content:
            text/csv:
              schema:
                type: string
            application/pdf:
              schema:
                type: string
                format: binary
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    MovementType:
      type: string
      enum: [stock_in, used_in_production, sold, damaged, missing_lost, returned]
...
    ReportType:
      type: string
      enum: [daily, weekly, missing, damaged, branch_activity]
...
    StockMovement:
      type: object
      properties:
        id:
          type: integer
        itemId:
          type: integer
        itemName:
          type: string
          nullable: true
        branchId:
          type: integer
        branchName:
          type: string
          nullable: true
        userId:
          type: integer
        userName:
          type: string
          nullable: true
        type:
          $ref: "#/components/schemas/MovementType"
        quantity:
          type: number
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - itemId
        - branchId
        - userId
        - type
        - quantity
        - createdAt
    CreateStockMovementRequest:
      type: object
      properties:
        itemId:
          type: integer
        branchId:
          type: integer
        type:
          $ref: "#/components/schemas/MovementType"
        quantity:
          type: number
        note:
          type: string
          nullable: true
      required:
        - itemId
        - branchId
        - type
        - quantity
...