# lib/api-zod/src/generated/types/movementType.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BakeryStock API specification
 * OpenAPI spec version: 0.1.0
 */

export type MovementType = typeof MovementType[keyof typeof MovementType];


export const MovementType = {
  stock_in: 'stock_in',
  used_in_production: 'used_in_production',
  sold: 'sold',
  damaged: 'damaged',
  missing_lost: 'missing_lost',
  returned: 'returned',
} as const;