# lib/api-zod/src/generated/types/stockMovement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BakeryStock API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MovementType } from './movementType';

export interface StockMovement {
  id: number;
  itemId: number;
  itemName?: string | null;
  branchId: number;
  branchName?: string | null;
  userId: number;
  userName?: string | null;
  type: MovementType;
  quantity: number;
  note?: string | null;
  createdAt: Date;
}